s3-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
boto3==1.34.34
pyjwt==2.8.0
python-multipart==0.0.6
requests==2.31.0